+++ test/browser/element-filter.spec.js
@@ -4,8 +4,8 @@ var expect = window.chai.expect;
 describe('ElementFilter', function () {
   describe('#filter', function () {
     it('filtering on table element.', function () {
-      var hit = ElementUtil.filter('table.filter', 'james');
-      expect(hit).to.equal(1);
+      var res = ElementUtil.filter('table.filter', 'james');
+      expect(res.filtered.length).to.equal(1);
     });
   });
 });