+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.6.0] - 2018-05-10
+
 ### Changed
 
 - Improved ElementFilter. #13
@@ -107,7 +109,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 First release. Please see [ElementUtil API](https://github.com/archco/element-util/tree/master/docs).
 
-[Unreleased]: https://github.com/archco/element-util/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/archco/element-util/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/archco/element-util/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/archco/element-util/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/archco/element-util/compare/v1.4.3...v1.5.0
 [1.4.3]: https://github.com/archco/element-util/compare/v1.4.0...v1.4.3